modify training example
