+++ examples/scripts/sentiment_tuning.py
@@ -16,10 +16,11 @@ from dataclasses import dataclass, field
 from typing import Optional
 
 import torch
+import tyro
 from datasets import load_dataset
 from peft import LoraConfig
 from tqdm import tqdm
-from transformers import AutoTokenizer, HfArgumentParser, pipeline
+from transformers import AutoTokenizer, pipeline
 
 from trl import AutoModelForCausalLMWithValueHead, AutoModelForSeq2SeqLMWithValueHead, PPOConfig, PPOTrainer, set_seed
 from trl.core import LengthSampler
@@ -30,79 +31,58 @@ tqdm.pandas()
 
 @dataclass
 class ScriptArguments:
-    """
-    The name of the Casual LM model we wish to fine with PPO
-    """
-
-    # NOTE: gpt2 models use Conv1D instead of Linear layers which are not yet supported in 8 bit mode
-    # models like gpt-neo* models are more suitable.
-    model_name: Optional[str] = field(default="lvwerra/gpt2-imdb", metadata={"help": "the model name"})
-    log_with: Optional[str] = field(default=None, metadata={"help": "use 'wandb' to log with wandb"})
-    learning_rate: Optional[float] = field(default=1.41e-5, metadata={"help": "the learning rate"})
-    mini_batch_size: Optional[int] = field(default=128, metadata={"help": "the PPO minibatch size"})
-    batch_size: Optional[int] = field(default=128, metadata={"help": "the batch size"})
-    gradient_accumulation_steps: Optional[int] = field(
-        default=1, metadata={"help": "the number of gradient accumulation steps"}
+    ppo_config: PPOConfig = field(
+        default_factory=lambda: PPOConfig(
+            model_name="lvwerra/gpt2-imdb",
+            query_dataset="imdb",
+            reward_model="sentiment-analysis:lvwerra/distilbert-imdb",
+            learning_rate=1.41e-5,
+            log_with=None,
+            mini_batch_size=128,
+            batch_size=128,
+            gradient_accumulation_steps=1,
+            early_stopping=False,
+            target_kl=6.0,
+            kl_penalty="kl",
+            seed=0,
+            use_score_scaling=False,
+            use_score_norm=False,
+            score_clip=None,
+        )
     )
-    early_stopping: Optional[bool] = field(default=False, metadata={"help": "whether to early stop"})
-    use_peft: Optional[bool] = field(default=False, metadata={"help": "whether to use peft"})
-    use_seq2seq: Optional[bool] = field(default=False, metadata={"help": "whether to use seq2seq models"})
-    kl_penalty: Optional[str] = field(
-        default="kl",
-        metadata={
-            "help": "kl penalty options: 'kl': model_logp - ref_logp,  'abs': abs(kl),  'mse': mean squared error mse(kl) and 'full': the actual kl for all tokens in the distribution"
-        },
+    query_dataset: str = field(default="imdb", metadata={"help": "the dataset to query"})
+    use_seq2seq: bool = field(default=False, metadata={"help": "whether to use seq2seq models"})
+    use_peft: bool = field(default=False, metadata={"help": "whether to use peft"})
+    peft_config: Optional[LoraConfig] = field(
+        default_factory=lambda: LoraConfig(
+            r=16,
+            lora_alpha=16,
+            bias="none",
+            task_type="CAUSAL_LM",
+        ),
     )
-    target_kl: Optional[float] = field(default=0.1, metadata={"help": "kl target for early stopping"})
-    seed: Optional[int] = field(default=0, metadata={"help": "the random seed"})
-    use_score_scaling: Optional[bool] = field(default=False, metadata={"help": "Use score scaling"})
-    use_score_norm: Optional[bool] = field(
-        default=False, metadata={"help": "Use score normalization. Only applicable if use_score_scaling is True"}
-    )
-    score_clip: Optional[float] = field(default=None, metadata={"help": "Score clipping"})
 
 
-parser = HfArgumentParser(ScriptArguments)
-script_args = parser.parse_args_into_dataclasses()[0]
+args = tyro.cli(ScriptArguments)
 
-config = PPOConfig(
-    model_name=script_args.model_name,
-    learning_rate=script_args.learning_rate,
-    log_with=script_args.log_with,
-    mini_batch_size=script_args.mini_batch_size,
-    batch_size=script_args.batch_size,
-    gradient_accumulation_steps=script_args.gradient_accumulation_steps,
-    early_stopping=script_args.early_stopping,
-    target_kl=script_args.target_kl,
-    kl_penalty=script_args.kl_penalty,
-    seed=script_args.seed,
-    use_score_scaling=script_args.use_score_scaling,
-    use_score_norm=script_args.use_score_norm,
-    score_clip=script_args.score_clip,
-)
-
-# set seed before initializing value head for deterministic eval
-set_seed(config.seed)
 
 # We then define the arguments to pass to the sentiment analysis pipeline.
 # We set `return_all_scores` to True to get the sentiment score for each token.
 sent_kwargs = {"return_all_scores": True, "function_to_apply": "none", "batch_size": 16}
 
-trl_model_class = (
-    AutoModelForCausalLMWithValueHead if not script_args.use_seq2seq else AutoModelForSeq2SeqLMWithValueHead
-)
+trl_model_class = AutoModelForCausalLMWithValueHead if not args.use_seq2seq else AutoModelForSeq2SeqLMWithValueHead
 
 
 # Below is an example function to build the dataset. In our case, we use the IMDB dataset
 # from the `datasets` library. One should customize this function to train the model on
 # its own dataset.
-def build_dataset(config, dataset_name="imdb", input_min_text_length=2, input_max_text_length=8):
+def build_dataset(config, query_dataset, input_min_text_length=2, input_max_text_length=8):
     """
     Build dataset for training. This builds the dataset from `load_dataset`, one should
     customize this function to train the model on its own dataset.
 
     Args:
-        dataset_name (`str`):
+        query_dataset (`str`):
             The name of the dataset to be loaded.
 
     Returns:
@@ -112,7 +92,7 @@ def build_dataset(config, dataset_name="imdb", input_min_text_length=2, input_ma
     tokenizer = AutoTokenizer.from_pretrained(config.model_name)
     tokenizer.pad_token = tokenizer.eos_token
     # load imdb with datasets
-    ds = load_dataset(dataset_name, split="train")
+    ds = load_dataset(query_dataset, split="train")
     ds = ds.rename_columns({"text": "review"})
     ds = ds.filter(lambda x: len(x["review"]) > 200, batched=False)
 
@@ -129,43 +109,41 @@ def build_dataset(config, dataset_name="imdb", input_min_text_length=2, input_ma
 
 
 # We retrieve the dataloader by calling the `build_dataset` function.
-dataset = build_dataset(config)
+dataset = build_dataset(args.ppo_config, args.query_dataset)
 
 
 def collator(data):
     return dict((key, [d[key] for d in data]) for key in data[0])
 
 
+# set seed before initializing value head for deterministic eval
+set_seed(args.ppo_config.seed)
+
 # Now let's build the model, the reference model, and the tokenizer.
-if not script_args.use_peft:
-    ref_model = trl_model_class.from_pretrained(config.model_name)
+if not args.use_peft:
+    ref_model = trl_model_class.from_pretrained(args.ppo_config.model_name)
     device_map = None
     peft_config = None
 else:
-    peft_config = LoraConfig(
-        r=16,
-        lora_alpha=16,
-        bias="none",
-        task_type="CAUSAL_LM",
-    )
+    peft_config = args.peft_config
     ref_model = None
     device_map = {"": 0}
 
 model = trl_model_class.from_pretrained(
-    config.model_name,
+    args.ppo_config.model_name,
     device_map=device_map,
     peft_config=peft_config,
 )
 
 
-tokenizer = AutoTokenizer.from_pretrained(config.model_name)
+tokenizer = AutoTokenizer.from_pretrained(args.ppo_config.model_name)
 
 # GPT-2 tokenizer has a pad token, but it is not eos_token by default. We need to set it to eos_token.
 # only for this model.
 tokenizer.pad_token = tokenizer.eos_token
 
 # We then build the PPOTrainer, passing the model, the reference model, the tokenizer
-ppo_trainer = PPOTrainer(config, model, ref_model, tokenizer, dataset=dataset, data_collator=collator)
+ppo_trainer = PPOTrainer(args.ppo_config, model, ref_model, tokenizer, dataset=dataset, data_collator=collator)
 
 # We then build the sentiment analysis pipeline, passing the model name and the
 # sentiment analysis pipeline arguments. Let's also make sure to set the device
